feat: create RouterBase and add common methods to it
To add support for route grouping we need another class
this class is called RouteRegistrar. This class holds the same methods
as the Router class. So, I decided to make RouterBase class that will
contain all common methods that are shared methods between Router and
RouteRegistrar is created in RouterBase.

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	new file:   packages/makex/src/Foundation/Routing/RouteRegistrar.ts
	modified:   packages/makex/src/Foundation/Routing/Router.ts
	new file:   packages/makex/src/Foundation/Routing/RouterBase.ts